Relax restriction on forward steps (#798)
For inference, remove requirement that n_forward_steps is a multiple of
n_forward_steps_in_memory. See Issue #664.

This PR makes changes to the inference data loader as well as the
TimeMeanAggregator (in the aggregator the change uses .sum instead of
.mean, see some discussion of this here:
https://github.com/ai2cm/full-model/issues/252 ) so that the total
number of steps doesn't need to be divisible by the number of forward
steps. In test_inference.py the test test_inference_writer_boundaries
was modified to include a case where the total number of steps is not
divisible by the number of forward steps. It also removes various checks
in the code where it requires total steps is divisible by forward steps.

---------


Co-authored-by: Finn Rebassoo <rebassoo@login38.chn.perlmutter.nersc.gov>
Co-authored-by: Jeremy McGibbon <mcgibbon@uw.edu>