Don't pass forward step args to run_inference functions (#859)
As shown in #798, we can determine `forward_steps_in_memory` from the
data returned by the inference loader, and so it is not necessary to
pass this explicitly to the `run_inference` and `run_data_inference`
functions. This PR stops doing that, and also determines the total
`n_forward_steps` from the provided data loaders. This reduces
likelihood of bugs due to user providing values for these arguments that
are inconsistent with the inference data loader also being provided.