Clean up diffusion config for downscaling models (#862)
The noise configuration for diffusion models can get a bit confusing.
This PR attempts to clean this configuration up in preparation for
implementing the sampling algorithm.

Changes:
- Hard codes `sigma_data` to 1.0 since our normalized data always has
standard deviation = 1.0 and removes this parameter from the
configuration or build methods throughout the codebase.
- Hard codes sigma_min and sigma_max to None for EDMPrecond since these
parameters are actually unused by the preconditioner. They are likely to
become configurable when we add the sampling code.