Set sigma coordinates to empty tensor if missing (#869)
While the main ACE project requires sigma coordinates, this is not true
of all potential XarrayDataset users, e.g. downscaling. This PR sets the
default value for sigma coordinates to the empty tensor,
torch.tensor([]), if they are missing from the dataset.

Changes:
- Instead of raising an exception if the dataset is missing ak and bk,
log a warning and set the ak, bk coordinates to the empty tensor.

---------

Co-authored-by: Oliver Watt-Meyer <oliverwatt@gmail.com>