Fix diffusion training for downscaling by using MSE instead of sum of SE (#886)
Due to the custom weighted loss function for diffusion, a mistake crept
in which performs sum instead of mean to compute MSE, as is done by
torch's MSELoss class. This results in a training objective which is not
properly conditioned. This PR fixes this issue by using mean instead of
sum to aggregate the squared errors across all pixels and examples in
the batch.