Create github-pages documentation (#888)
This PR updates the documentation build workflow so that the build docs
are deployed to the gh-pages branch, which has now been configured on
the repo. A link to [the
documentation](https://ai2cm.github.io/full-model/html/index.html) has
been added to the top-level README. Users must be logged in and have
repository access to see the documentation.

Github actions code comes from [this
example](https://coderefinery.github.io/documentation/gh_workflow/).

The resulting URL gets redirected to a fairly weird url, I don't know
how to avoid this other than to make an entire `ace-dev.github.io`
repository to host the docs as suggested in the github pages
documentation. We could perhaps do this for the public release, though
we may not use github-pages for that release at all.

Required for #751 .