Add a script to rechunk FME datasets using xarray_beam (#891)
Now that Zarr is a compatible format for training and inference, we have
been working on trying to find a chunksize that optimizes for both
cases. This PR adds a processing script that makes it easy to move a
dataset to a new time chunk size. It uses xarray_beam with Google
Dataflow to provide a quick and cost-effective method to adjust our
datasets.

A run script is provided `run-rechunk.sh` that was used to submit
Dataflow jobs for each ensemble member in an FME dataset.