Change default output format to NETCDF4 (#900)
The reason this script used the NETCDF3_64BIT default is that we used to
use `open_mfdataset` in our data loader, and this function has a weird
memory leak when using NETCDF4 files. But we no longer use that
function, so we should just use NETCDF4. I found this led to a
significant speedup (ballpark 3-4x) in the netCDF conversion script.