Add more geopotential pressure levels to ERA5 dataset (#914)
These variables will be useful for training a DLWP baseline. cc
@anna-dodson

I also increase the output chunk size, since we are not training
direclty on zarr for now, and this larger chunk size makes the dataset
easier to work with (e.g. computing statistics over the full dataset, or
copying/deleting the dataset).