Spectral power metric and aggregator (#923)
The power spectrum is a typical diagnostic to assess the smoothness of
ML weather/climate emulators. This add this diagnostic to our standard
inference metrics. It is accumulated over all forward steps for
inline/standalone inference. Could also add this for 1-step validation,
but I've left that out for now.

Changes:
- add `SphericalPowerSpectrumAggregator` and
`PairedSphericalPowerSpectrumAggregator` and use the latter in inline
and standalone inference

- [x] Tests added