Add gc.collect to start of each epoch to resolve CUDA errors (#942)
Similar to what was described
[here](https://github.com/ai2cm/full-model/issues/740#issuecomment-2079765293),
got an odd cuda error after adding the spherical power spectrum metric.
I tried resolving this the same it was done in that case (add a
`gc.collect()` after closing figure) but that didn't work. As suggested
in [this
issue](https://github.com/pytorch/pytorch/issues/67978#issuecomment-1661986812),
I added a `gc.collect()` call at the start of each epoch, and this seems
to have resolved the problem.