Add env var for using MPS if available (#950)
Add the ability to use MPS without code changes. I thought about
defaulting to MPS (over CPU) if it is available, but a lot of our tests
fail with that backend. The backend seems to not be mature enough to
support training, although inference does work in my experience.

Changes:
- `fme.get_device()` will return an MPS device if `FME_USE_MPS` is set
to 1 and MPS is available.
- Entrypoints log the current device at init.