Log ENSO index coefficients to wandb (#925)
Adds a set of wandb metrics that describe the regression coefficients of
predicted and target variables with a pre-computed ENSO index (1940-2020
Nino3.4). Produces maps of the regression coefficients for target,
prediction, and their difference, and scalar global-mean difference
magnitude of the coefficients, for each predicted variable. Also includes the
script used to pre-compute the Nino3.4 index.

Changes:
- No changes to public API.
- The ENSO aggregator will report metrics for inference longer than ~5
years, assuming inference overlap with the 1940-2020 Nino3.4 period. No
metrics are reported if there is not near-full overlap.