Implement residual diffusion (e.g. CorrDiff) for downscaling (#1002)
This PR implements the residual model for diffusion models in
downscaling. During training the idea is to subtract the base prediction
from the latent variable and the target. At generation time, we
essentially generate the residual using the iterative diffusion
algorithm and, as a final step, add in base prediction. This PR assumes
that the base prediction is simply an unparameterized (bicubic)
interpolation.

- [x] Tests added