HEALPix integration with ACE stepper code  (#1065)
Adds a new parameter to test_train_and_inference_inline, which tests the
HEALPixRecUNet integration with the stepper and pipeline.

We eliminate the time dimension for now in favor of moving all time
dimensioning logic to the stepper. As such, we now have inputs to the
model in the form [B, F, C, H, W]. This greatly simplifies the forward
pass of the model, which now only folds inputs, encodes, decodes,
unfolds, and concats the appropriate prognostics and output-only
(added).

I also add area_weights to be Optional, and whenever it's None, don't
perform aggregation. In cases where metrics do not exist, I assert that
the paths are not present. Prediction doesn't happen during this test,
as well, since DataWriter is not configured for healpix. This changes a
number of API signatures for places we use HEALPix in aggregation.

def load_stepper(
    checkpoint_file: str,
  -  area: torch.Tensor,
  +  area: Optional[torch.Tensor],
    sigma_coordinates: SigmaCoordinates,
    ocean_config: Optional[OceanConfig] = None,
) -> SingleModuleStepper:

Data writer save initial condition/ now also takes an additional
argument, where we leave the work of writing to later: Handle writing
HEALPix data https://github.com/ai2cm/full-model/issues/1089.
def save_initial_condition(
        self,
        ic_data: Dict[str, torch.Tensor],
        ic_time: xr.DataArray,
        snapshot_dims: List[str],
    ):

---------

Co-authored-by: Anna Dodson <annad@climate-cirrascale-72.reviz.ai2.in>