Add --very-fast test flag, enforce timeout (#1104)
This PR adds a `--very-fast` pytest flag which skips all tests longer
than 5 seconds (and a couple in the 2-5s range). It also enforces a 5s
timeout for these tests, a 30s timeout for `--fast` tests, and a 60s
timeout by default. This is meant to quickly run tests unrelated to the
change you're making when iterating locally, rather than to provide full
coverage of the code (which `--fast` is meant to do, mostly).

Changes:
- Added `--very-fast` pytest flag
- Added timeouts for pytest tests, 60s by default, 30s if using
`--fast`, and 5s if using `--very-fast`

- [ ] Tests added