Fix --check-untyped-defs typing issues in tests (#1123)
This PR fixes a number of typing issues identified by mypy if run with
`--check-untyped-defs` in test files.

I found and applied these fixes by adding `--check-untyped-defs` to the
mypy init in pre-commit-config and running pre-commit --all-files
repeatedly. It can't be enabled yet because there are additional
failures in non-test files.

Needed for #1098.