Use serialized vertical coordinate (#1164)
This means we don't have to read the vertical coordinate from the
forcing data, which makes code simpler and makes generation of new
forcing datasets more straightforward.

It was done this previously when we didn't serialize the vertical
coordinate. But even our publicly released checkpoint from last November
includes the necessary info, so I don't think backwards compatibility is
an issue here. I did include a cast to single precision, since the distributed
ACE checkpoint serialized with double precision (not the case for
more recent checkpoints).

Resolves #1150