Near-zero global-mean ENSO coefficients (#1198)
Tweaks the ENSO regression coefficient aggregator's computations to
ensure that the resulting maps don't have large global-mean offsets due
to an artifact in computation, by moving the ENSO index mean removal to
after reindexing to the set of evaluation timesteps.

Some non-zero global- or regional-means that are not really ENSO-related
may remain in the coefficient maps if e.g. there is a global- or
regional-mean long-term trend in a variable and the Nino3.4 index also
has a long-term trend (we don't detrend the index prior to computing
ENSO correlation).