Use num_data_workers=0 in tests (#1222)
On my Mac I found this sped up the integration tests a lot. It seems
that initializing parallel data workers (even one of them!) is very slow
on Mac. Specifically, running all the tests in `fme/ace/inference` went
from 128s to about 19s while running tests in `fme/ace/test_train.py`
when from 4m4s to 1m37s.

On CI (running on linux...) it is not clear if this makes a robust
difference. Initially I thought it sped the CPU tests up by about 30s,
but that is within range of run-to-run variability.