Add RepeatedInterval subsetting option for datasets (#1213)
This PR adds a new subset selection class, `RepeatedInterval`,allowing
for a repeated interval selection of samples in a Dataset. This will be
useful for using a single dataset with the downscaling data and, for
example, using 6-days of every week for training while leaving the 7th
for validation.

Changes:
- `XarrayDataConfig` now accepts a `RepeatedInterval` as a subset type
```
    Examples:
        To return values from the first 3 items of every 6 items, use:
        >>> RepeatedInterval(interval_length=3, repeat=6, start=0)
        To return a days worth of values starting after 2 days from every 7-day
        block, use:
        >>> RepeatedInterval(interval_length="1d", repeat="7d", start="2d")
    """
```
- `subset_dataset` is now a member of an `XarrayDataset` instead of a
separate function call. I made this adjustment since that function needs
some internal knowledge of properties to transfer anyway
- Added a new helper method, `transfer_properties` to transfer relevant
properties between torch `Dataset` objects that also allows for
specified overrides
- the abstract `Dataset` type definition now includes a
`BASE_PROPERTIES` attribute that will be used for default transfers

- [x] Tests added