Refactor training to rely on protocol config class (#1224)
This PR refactors `Trainer` so we can use dependency injection. This
will allow us to change the way data is loaded or stepping is done
without needing to put the code in `core`. It also makes more explicit
some of the coupling between the `TrainConfig` class and the Trainer.

`TrainConfigProtocol` could be further reduced in complexity if we move
some of its responsibilities onto `Trainer`, like the configuration and
checkpointing logic methods. This would reduce duplication when we add
other configuration classes.

Changes:
- Instead of taking `TrainConfig` and building everything inside
`Trainer.__init__`, `Trainer` now takes in built components which are
built inside a `build_trainer` helper function, and requires only a
(much more) minimal protocol class for `config`.
- Added a CheckpointPaths helper object to produce sub-paths of
`checkpoint_dir` corresponding to checkpoint locations.
- Added top-level methods to TrainConfig to retrieve nested sub-methods,
so that the protocol class need not be nested.

- [ ] Tests added