Add entrypoint for running segmented inference (#1147)
Update the `fme.ace.inference` entrypoint to allow repeating the
inference over an integer number of segments.

For example, this can be done with `python -m fme.ace.inference
config.yaml --segments 3`

The advantage of this is that it would allow automatic restarting of
inference from the last complete segment and makes preemption/restarting
much simpler. It also allows splitting of output datasets across
segments, which is useful to avoid large netCDF files. A downside of
using this approach is that inference runs would be split across
multiple wandb runs. Also, the aggregator outputs would be saved for
each segment and it would be up to the user to combine them if
necessary.

---------

Co-authored-by: Jeremy McGibbon <mcgibbon@uw.edu>