Fix days since reference calculation for intervals longer than 292 years (#1235)
This change was made so that inference can go past 292 years. Without
this fix, you run into an error that the time difference exceeds range
of values that can be expressed at the nanosecond resolution. After this
fix was implemented it was tested that it works for a 500 year
inference.

---------

Co-authored-by: Finn Rebassoo <rebassoo@login21.chn.perlmutter.nersc.gov>
Co-authored-by: Finn Rebassoo <rebassoo@login14.chn.perlmutter.nersc.gov>
Co-authored-by: Finn Rebassoo <rebassoo@login09.chn.perlmutter.nersc.gov>