Improve timing categories and fix Looper docstring (#1254)
Currently we use `"run_on_batch"` as the name of the timer category that
encompasses the forward prediction of the ML model during inference.
This is a bit of a misnomer, given that we do not use the `run_on_batch`
method in inference. This PR updates that term to be called
`"forward_prediction"` and splits off a separate category for computing
derived variables. It also splits the writer and aggregators timings
into two.

In summary, categories after this PR are:
`inference` (which is a total duration)
`initialization`
`data_loading`
`forward_prediction`
`compute_derived_variables`
`data_writer`
`aggregator`
`wandb_logging`
`final_writer_flush`