Masking generated output (#1279)
This PR adds the ability to apply masking to generated output based on
the config. It is not actually used yet but we will need it for ocean
emulator. We are making the masking feature generic so that it can be
used for other future applications.

Changes:
- Added `fme.core.masking`: contains Masking and MaskingConfig
- `ClimateData`: added `all_level_keys` which returns a list of all
vertical level keys under a specified standard name

- [x] Tests added