Refactor inference aggregators to use BatchData (#1282)
This PR refactors the inference aggregators to use BatchData as their
input type. This will make it possible to make the aggregators generic
over this type within the inference code.

Changes:
- Inference aggregators now take two BatchData as argument instead of
one TrainOutput. Training and validation aggregators still use
TrainOutput as they measure what happens during a training call.
- Aggregator ABCs have moved to `fme.core.generics.aggregator`.
- InferenceAggregator now takes in unused `target` and `normalize`
arguments to conform to the same API as the InferenceEvaluatorAggregator
- StandardNormalizer.normalize now takes in the more permissive
TensorMapping type so it can be used on BatchData (which contains
TensorMapping).

- [ ] Tests added