Add some analysis packages to `fme` conda environment (#1304)
It is convenient to be able to use the same environment for developing
`fme`, running ACE and doing analysis on ACE outputs. This PR adds a
`analysis-deps.txt` file to the root of full-model, which is used during
the `make create_environment` rule (but these things are not added as
requirements to `fme`).