Add scripts to strip optimization state from 3 ACE models (#1337)
Convenience script to prepare three ACE checkpoints for upload to
Hugging Face. We will host the models without the optimizer states so
that they are smaller. We could separately include the optimizer state
in the Hugging Face model repos also, but this doesn't seem high
priority. I also convert the area/sigma coordinates to single precision
(currently the old ACE checkpoint has them saved in double precision).
Although we already do this in FME code when deserializing the model,
this is a bit trickier on MPS which doesn't even support double
precision, so figured we might as well just distribute the checkpoints
with single precision coordinates.