Remove n_forward_steps attribute from dataset (#1355)
This PR removes the `n_forward_steps` attribute from dataset objects.
This was only used for logging during the data comparison run (not
during inference), and heavily uses dynamic typing. The attribute isn't
transferred by `transfer_properties`, for example. It is also a source
of confusion that this isn't referring to the `n_forward_steps` of a
batch as specified in training config.

Changes:
- Removed `n_forward_steps` attribute from dataset objects
- `n_forward_steps` is no longer logged during coupled data loading or
data comparison.

- [ ] Tests added