Add InferenceGriddedData and new requirements types (#1365)
This PR adds InferenceDataABC and InferenceGriddedData so that an
initial condition can be retrieved directly from the inference dataset
for inference. This has the benefit of reducing the API of the dataset
passed to inference.

Changes:
- Added InferenceGriddedData class
- Renamed DataRequirements to WindowDataRequirements
- Added PrognosticStateDataRequirements class, now required for
`get_inference_data`
- `get_inference_data` now returns both an initial condition and a
DataLoader of forward windows. This increases our memory requirements
during training to include an additional prognostic state kept when not
doing inference, but this should be negligible compared to memory
requirements for gradients.
- SingleModuleStepperConfig now provides
`get_evaluation_window_data_requirements`,
`get_prognostic_state_data_requirements`, and
`get_forcing_window_data_requirements` methods.
- Deleted unused `get_data_requirements` method from
`SingleModuleStepper`
- `run_inference` and `run_inference_evaluator` now take in the initial
condition and a loader of reference data, instead of a data object
containing both
- Requirements during inference now show the size of a window of data
instead of the whole data, consistent with what is done during training.
- `get_forcing_data` now takes in an initial condition and returns an
object with an initial condition, to guarantee alignment between the
initial condition timesteps and the loaded forcing data.

- [x] Tests added