Stop pinning uv in actions/Makefile (#1377)
Resolves #875 

I ran `make create_environment` locally and it worked fine.

For clarity, I also updated the `uv pip install -e` calls to use `./fme`
to make it explicit that we want to install the local directory and not
the `fme` pyPI package. Not strictly necessary because this is assumed
when using `-e` flag, but it gives us a bit of safety if pip or uv
change behavior again.