Add configurations for slab ocean datasets (#992)
This PR adds configurations for slab ocean datasets at 4° and 1° resolution.  Here we compute stats using the first four ensemble members in the 1xCO2, 2xCO2, and 4xCO2 climates, and exclude data from the fifth ensemble member in each of those climates, as well as all ensemble members in the 3xCO2 climate.  Processing configurations are also provided for non-equilibrium climate cases.

---------

Co-authored-by: Brian Henn <brianhenn@allenai.org>