Use common function for writing initial condition and restart (#1414)
Refactors `DataWriter.save_initial_condition` function to a more general
`.write` function so that it can be used for saving both the initial
condition and the restart. Lets us delete a bunch of code, and also now
restarts should work for steppers where `n_ic_timesteps` is greater than
1.

The only change in behavior is that `run_dataset_comparison` no longer
saves a restart file, which seems fine. That method already does not
save the initial condition, so this just makes it consistent with that.

Changes:
- refactor `DataWriter.save_initial_condition` to `DataWriter.write`
- delete `RestartWriter` and `PairedRestartWriter`