Resume wandb on id instead of folder (#1440)
Changes wandb resumption strategy from relying on the wandb `dir` being
the experiment directory to instead resuming with a wandb run id string,
which is now saved in the experiment directory. The wandb `dir` is set
to a temporary location, which is the preferred approach, so the manual
removal of the wandb files is no longer needed and the experiment directory
is much less cluttered. Adds `wandb.init`-like behavior to `MockWandB` for testing.