Fix type enforcement in Registry by validating against the generic base type (#1476)
This PR addresses a minor bug in the Registry class where the register
method did not enforce type constraints for registered classes. The
previous implementation allowed any class to be registered, even if it
was not a subclass of the base type specified by the generic parameter
T.

Changes:
- Introduced logic to dynamically retrieve the base type (T) using the
__orig_class__ attribute of the instance.
- Added a check using issubclass() to ensure that any class being
registered is a subclass of the inferred base type. Raised a TypeError
when a class that does not inherit from the base type is registered.
Without this, we can only spot this error through static type checking.
- Simplified the type annotations by removing the unused TT TypeVar,
which was redundant in the context of the register method.
- Minor fixs in module.py, perturbation.py and corrector.py to ensure
consistency wrt the changes in register() typing.
- Added tests to verify the behavior of register with and without
specified parameter T.

- [x] Tests added

---------

Co-authored-by: Jeremy McGibbon <mcgibbon@uw.edu>