Formalize overriding Stepper configuration options at inference time (#1468)
This PR moves all options that can be overridden in the Stepper at
inference time into a `StepperOverrideConfig` passed to either the
`InferenceConfig` or `InferenceEvaluatorConfig`. The `ocean` option has
been removed. In so doing, we have also implemented the ability to
override the multi-call configuration used for inference.

Changes:
- Added a `stepper_override` option to `InferenceConfig` and
`InferenceEvaluatorConfig`, which can be set to `None` or a
`StepperOverrideConfig` object.
- Removed the `ocean` option in `InferenceConfig` and
`InferenceEvaluatorConfig`, since it has been moved to the
`StepperOverrideConfig`.
- Enabled overriding the multi-call configuration of the stepper used
for inference through the `StepperOverrideConfig`.
- Moved `load_stepper_config` and `load_stepper` functions to
`stepper.py` as they may be useful in other contexts.

- [x] Tests added

Resolves #1452