Add config override to fme.ace.train via CLI dotlist (#1467)
Allows overriding of individual elements of config. This can be useful
to avoid having to copy and paste yaml configs, or having to use yq or
other tools to update yaml. However, as far as I can tell it is not
possible to update lists using the dotlist, so not all parts of our yaml
configs can be updated with this method.

Changes:
- add `omegaconf` dependency. We could also implement our own "dotlist"
parser, but it's nice to avoid this. `omegaconf` is pretty widely used
(e.g. in hydra, and in NVIDIA's modulus) but also is not being very
actively developed at the moment
https://github.com/omry/omegaconf/issues/1200.
- `fme.ace.train` entrypoint now accepts an optional `--override`
argument, for example `python -m fme.ace.train config.yaml --override
n_forward_steps=3`
- `fme.ace.validate_config` similarly takes an optional `--override`
argument

- [x] Tests added

Resolves #1432

---------

Co-authored-by: Jeremy McGibbon <mcgibbon@uw.edu>