Check for `requires_grad` before DDP wrap (#1482)
If all `SingleModuleStepper` module parameters are frozen, wrapping the
module with `DistributedDataParallel` results in an error. As a result,
it was not possible to, for example, use a fully fixed atmosphere
component stepper in `CoupledStepper`. With this PR, we first check for
learnable parameters before the DDP wrap.

Changes:
- Module is wrapped in `DummyModule` when no parameter requires grad,
regardless of the result of `Distributed.is_distributed()`.