Multisource option combines horizontal datasets  (#1485)
Sometimes we would like to load datasets from a different source. For
example, we are training an ocean emulator, but we want to loa surface
variables from the atmosphere data. In the past, the only way to do this
is to re-process the data and add those variables to the same
netcdf/zarr file. Another use case is documented in #1125. This PR
allows the option to combine different datasets horizontally. For now,
we expect stats to have every variables across different sources (i.e.,
stats need to be combined manually).

Changes:
- `DataLoaderConfig` now also accepts a dictionary of datasets where
data from different sources will be concatenated as long as the time
index are identical
  ```
  dataset:
    source1:
      - data_path: ...
    source2:
      - data_path: ...
  ```
Datasets within the dictionary will be merged, while the datasets in the
list will be concatenated.
- merged dataset is also an option in initial condition config and
forcing data loader config
- `MergedXarrayDataset` is a new wrapper class around `XarrayDataset`
which returns all variables in the different dataset sources


- [x] Tests added

Resolves #1125