Enable adding constant residual term to energy budget (#1507)
Sometimes our target data has a mean residual in the total energy
budget. In this case it is useful to account for this in our energy
budget correction.

Changes:
- added new `EnergyBudgetConfig` with a new
`constant_unaccounted_heating` config option
- deprecated existing method of activating the energy corrector
- the `net_energy_flux_into_atmosphere` variable now includes this
"unaccounted heating" term
- the "unaccounted heating" is also added to the derived variables. In
case of not imposing the energy corrector, or for the target data, it
will be equal to zero.

- [x] Tests added