Fill nans for sea ice fraction before regridding (#1540)
Previously, we did not fill nans in sea ice fraction before regridding.
So nan values are ignored during regridding and the weights are
renormalized. But we did fill nans for ocean fraction, which led to an
inconsistency where sea ice fraction did not consider land points but
ocean fraction did. This PR addresses this issue.