Corrector and derive func build methods on coordinate (#1546)
This PR resolves static typing issues caused by `VerticalCoordinate` by
replacing it with `StepperVerticalCoordinate` containing build methods
for the corrector and derive function. This reflects the current state
of our code, which determines which type to build based on data on disk.

Changes:
- Added ABC `StepperVerticalCoordinate` with new methods
`build_corrector` and `build_derive_function` to be used by the stepper
- Added `HasAtmosphereVerticalIntegral` protocol class to be used in
metric and corrector code which is not allowed to depend on the
coordinate module to avoid circular dependencies. This also meant using
`Optional` instead of NullVerticalCoordinate in these parts of the code.
- Removed unused `vertical_coordinate` arguments in aggregators. These
can be added to aggregators that do use them if and when they are
introduced - the code may be quite different by then and removing them
avoids awkwardly typed code for now.

- [ ] Tests added

Resolves #1544

---------

Co-authored-by: Oliver Watt-Meyer <oliverwatt@gmail.com>