Log Nino3.4 index (#1528)
Adds a new wandb metric (plotly image `inference/enso_index/nino34_index`), which is a dynamically computed Nino3.4 index based on the `surface_temperature` of the predictions and target. The computation uses canonical Nino3.4 methods and reports with a monthly time coordinate; it is a 5-month lagging mean, so the first 4 months of a simulation it will be null. It is computed without a "climate-invariant" correction to avoid needing the ocean fraction forcing data variable. In addition, a netCDF of the index is saved for standalone cases ("enso_index_diagnostics.nc").