Fix: test_train with multiple years of data (#1549)
The comments in `test_train.py` indicate that the intent is to run tests
with inline inference for >2 years (this would mean that the annual
aggregators, etc. are exercised). But in practice this did not happen
because the configuration of number of timesteps ignores the
`timestep_days` input, and so it ran for about two months instead; this
PR fixes that.