Require layer norm for NoiseConditionedSFNO, use global_layer_norm (#1574)
The `normalization_layer` option is currently not used and ignored, but
only layer_norm is supported.

`global_layer_norm` was never put into the configuration class for
NoiseConditionedSFNO, and the default is to do column-local
normalization. We expect better results from global instance
normalization, and should test it.

Changes:
- Set default `normalization_layer` for NoiseConditionedSFNOBuilder to
`layer_norm`, and added post-init check that it is set to this value.
- Add `global_layer_norm` option to NoiseConditionedSFNOBuilder.

- [ ] Tests added