Refactor masking to require static mask tensor on init (#1579)
With https://github.com/ai2cm/full-model/pull/1558 the 3D ocean mask is
now loaded as part of the ocean `DepthCoordinate`, meaning that we have
access to the mask when building `OceanCorrector`. This PR refactors
`StaticMasking` (previously called `Masking`) so that we can directly
initialize it with the `DepthCoordinate.mask`. As a result, the `mask_*`
variables are no longer required to be ML forcing inputs in order to
mask out the land when emulating ocean fields.

Changes:
- Renames `Masking{Config}` to `StaticMasking{Config}`
- Removed the `mask_data` argument from `StaticMasking.__call__`.
- `StaticMaskingConfig.build` now has args `mask_2d` and `mask_3d`.
- Removed the `StaticMaskingConfig` attributes `mask_name` and
`surface_mask_name`.
- Adds `get_mask` and `get_idepth` methods to `DepthCoordinate`
- `OceanCorrector` has new init args `gridded_operations`,
`vertical_coordinate` and `timestep`.

- [x] Tests added

Resolves https://github.com/ai2cm/full-model/issues/1555