Add a climsst baseline configuration (#1573)
Our existing ace2-era5 baseline involves training the model on
historical CO2 and SST forcings, and then testing on an out-of-sample
inference period. There are some signs we may be overfitting in this
problem, such as behavior over-indexing on CO2, and in some cases seeing
the upper atmosphere humidity and temperature validation RMSEs go up
with training. We also see that in-sample inference loss goes down with
more training for configurations where out-of-sample inference loss goes
up. In other words, features which make the model more powerful tend to
worsen our inference skill.

A main goal we have right now is to align our training objective and
long-term inference skill. In particular, I am working on changes to our
model and training objective which might align model training with
improvements to long-term inference. Even on in-sample forcings, our
model has issues with long-term and short-term inference behavior which
we want to rectify, such as muted power spectra or long-term upper
atmospheric drifts.

The in vs out of sample setup of ace2-era5 makes it difficult to draw
conclusions for model changes like CRPS loss or diffusion modelling,
particularly when inference loss increases due to model changes. For
this reason I'll be evaluating these changes on our older "climSST"
data, where we believe we can still reproduce (and thus correct) these
issues.

The configuration here is based mainly on our latest ace2-era5
configuration, removing inputs and outputs that don't exist in that
dataset. I have also modified the configuration to use the "layer_norm"
option and gradient accumulation, which are changes we plan to
incorporate into the ace2-era5 baseline.

The inference start indices are selected to sample the annual and
diurnal cycles, while still being near the start of the inference period
to allow longer inline inference (the validation dataset is limited to
10 years).

Changes:
- Added `climsst` baseline configuration