Check for non-positive input surface pressure in energy corrector (#1601)
When training models with the energy corrector and multiple steps in the
loss, we encountered situations where training would crash in the first
epoch due to NaN loss. This happened to be for the same reason we
encountered occasional NaN loss in single step training—the surface
pressure somewhere on the globe is less than zero, which leads to
negative interface pressures, and NaNs when computing the geopotential
height required for computing the total energy path (taking the `log` of
negative values produces NaNs in [this
line](https://github.com/ai2cm/full-model/blob/753ae66826cff7aa51ce38756c05c3120c2674c2/fme/core/atmosphere_data.py#L376)).

The difference in multi-step training is that negative surface pressure
can appear in either the input or output of ACE, and during energy
correction we need to compute the total energy path for each. The
existing code only checks for non-positive surface pressure in the
output. This PR updates the code to also check for non-positive surface
pressure in the input.

Changes:
- We now check the surface pressure in both the input and output for
non-positive values before proceeding with the energy corrector.