Account for frozen precip in derived sfc energy flux (#1629)
Make the surface energy flux derived variable consistent with what is
used in the net energy flux into atmosphere (and in the total energy
corrector). Note that we still use the surface flux without frozen
precipitation in the slab ocean, which was done for consistency with
SHiELD. We could add a special derived variable for the version
explicitly without frozen precip, but for now I'm leaving that out.