Separate `.step()` responsibility into new object (#1605)
This PR defines a class `SingleModuleStep` which provides `.step()` as
in the current SingleModuleStepper. It refactors SingleModuleStepper to
defer to this class for `.step()`. This will allow later refactors to
provide other implementations of `.step()`.

This change is intentionally done without changing public methods or
features of SingleModuleStepper. The init method is changed, but we only
use the init methodprivately by constructor routines.

Changes:
- Define new SingleModuleStep class
- Use SingleModuleStep argument to SingleModuleStepper init

- [ ] Tests added