Ensure energy corrector doesn't clobber prognostic variables (#1648)
Resolves #1638. This is needed since "surface temperature" is a edge
case in our AMIP-style prognostic runs, i.e. it is a prognostic variable
but also used as a forcing variable over ocean, and so exists in the
"forcing data" provided to the corrector. A better long-term solution
might be defining separate `sea_surface_temperature`,
`land_surface_temperature` and `ice_surface_temperature` variables so we
would have a clearer separation between forcing/prognostic variables.